+++ src/bump_string.rs
@@ -210,14 +210,14 @@ where
     /// length of the string.
     #[must_use]
     #[inline(always)]
-    pub fn len(&self) -> usize {
+    pub const fn len(&self) -> usize {
         self.fixed.len()
     }
 
     /// Returns `true` if this `BumpString` has a length of zero, and `false` otherwise.
     #[must_use]
     #[inline(always)]
-    pub fn is_empty(&self) -> bool {
+    pub const fn is_empty(&self) -> bool {
         self.fixed.is_empty()
     }
 
@@ -703,7 +703,7 @@ where
     }
 
     unsafe fn insert_bytes<B: ErrorBehavior>(&mut self, idx: usize, bytes: &[u8]) -> Result<(), B> {
-        let vec = unsafe { self.as_mut_vec() };
+        let vec = self.as_mut_vec();
 
         let len = vec.len();
         let amt = bytes.len();