+++ src/fixed_bump_string.rs
@@ -83,14 +83,14 @@ impl<'a> FixedBumpString<'a> {
     /// length of the string.
     #[must_use]
     #[inline(always)]
-    pub fn len(&self) -> usize {
+    pub const fn len(&self) -> usize {
         self.initialized.len()
     }
 
     /// Returns `true` if this `FixedBumpString` has a length of zero, and `false` otherwise.
     #[must_use]
     #[inline(always)]
-    pub fn is_empty(&self) -> bool {
+    pub const fn is_empty(&self) -> bool {
         self.initialized.is_empty()
     }
 
@@ -567,7 +567,7 @@ impl<'a> FixedBumpString<'a> {
     }
 
     unsafe fn insert_bytes<B: ErrorBehavior>(&mut self, idx: usize, bytes: &[u8]) -> Result<(), B> {
-        let vec = unsafe { self.as_mut_vec() };
+        let vec = self.as_mut_vec();
 
         let len = vec.len();
         let amt = bytes.len();