+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **added:** missing string methods to `BumpBox<str>`: `len`, `is_empty`, `set_len`, `retain`, `clear`, `as_ptr`, `as_mut_ptr`, `remove`, `as_mut_bytes`
 - **added:** `Default` for `FixedBumpString`
 - **added:** string methods `pop`, `truncate`
+- **added:** made more `len` and `is_empty` methods `const`
 
 ## 0.10.1 (2024-10-14)
 - **added:** `MutBumpVecRev::{ append, into_flattened, unchecked_push(_with), as_non_null_{ptr, slice} }`